Optimize player huddling temperature modifier Add seasons card timing configs Add dimension base temperature Add config ranges

+++ src/main/java/sfiomn/legendarysurvivaloverhaul/client/tooltips/HydrationTooltip.java
@@ -51,7 +51,7 @@ public class HydrationTooltip {
     }
 
     public HydrationTooltip(HydrationEnum hydrationEnum) {
-        this(hydrationEnum.getHydration(), hydrationEnum.getSaturation(), hydrationEnum.getDirtiness());
+        this(hydrationEnum.getHydration(), (float) hydrationEnum.getSaturation(), (float) hydrationEnum.getDirtiness());
     }
 
     public String getPlaceholderTooltip() {
